requirements.txt
discord.py>=2.3.2
matplotlib>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.25.0
sentry-sdk>=1.39.0
Flask>=2.3.0
